update: added channels for server shutdown
