+++ file-transfer/main.go
@@ -4,8 +4,10 @@ import (
 	"context"
 	"file-transfer/app"
 	"flag"
-	// "os"
-	// "os/signal"
+	"os"
+	"os/signal"
+	"syscall"
+
 	"time"
 )
 
@@ -16,11 +18,25 @@ func main() {
 	a := app.App{}
 	a.Initialize()
 
+	done := make(chan bool)
+	quit := make(chan os.Signal, 1)
+	signal.Notify(quit, syscall.SIGINT, syscall.SIGTERM)
+
+	go func() {
+		<-quit
+
+		ctx, cancel := context.WithTimeout(context.Background(), 30*time.Second)
+		defer cancel()
+
+		if err := a.Close(ctx); err != nil {
+			panic(err)
+		}
+		close(done)
+	}()
+
 	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
 	defer cancel()
 
 	a.Run(&ctx, ":"+*port)
-	if err := a.Close(ctx); err != nil {
-		panic(err)
-	}
+	<-done
 }
